Rename internal weakref package to luagc
It is not just about managing weakrefs but also about managing finalization
of lua values.
