+++ runtime/internal/luagc/defaultpool_safepool.go
@@ -1,0 +1,8 @@
+//go:build safepool
+
+package luagc
+
+// NewPool returns a new Pool with an appropriate implementation.
+func NewDefaultPool() Pool {
+	return NewUnsafePool()
+}
