--- runtime/internal/weakref/defaultpool_safepool.go
@@ -1,8 +1,0 @@
-//go:build safepool
-
-package weakref
-
-// NewPool returns a new WeakRefPool with an appropriate implementation.
-func NewDefaultPool() Pool {
-	return NewUnsafePool()
-}
