+++ runtime/runtimecontextmanager.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/arnodel/golua/runtime/internal/weakref"
+	"github.com/arnodel/golua/runtime/internal/luagc"
 )
 
 const QuotasAvailable = true
@@ -40,7 +40,7 @@ type runtimeContextManager struct {
 	startTime        uint64
 	nextCpuThreshold uint64
 
-	weakRefPool weakref.Pool
+	weakRefPool luagc.Pool
 	gcPolicy    GCPolicy
 }
 
@@ -48,7 +48,7 @@ var _ RuntimeContext = (*runtimeContextManager)(nil)
 
 func (m *runtimeContextManager) initRoot() {
 	m.gcPolicy = IsolateGCPolicy
-	m.weakRefPool = weakref.NewDefaultPool()
+	m.weakRefPool = luagc.NewDefaultPool()
 }
 
 func (m *runtimeContextManager) HardLimits() RuntimeResources {
@@ -129,7 +129,7 @@ func (m *runtimeContextManager) PushContext(ctx RuntimeContextDef) {
 	m.messageHandler = ctx.MessageHandler
 	m.parent = &parent
 	if ctx.GCPolicy == IsolateGCPolicy || ctx.HardLimits.Millis > 0 || ctx.HardLimits.Cpu > 0 || ctx.HardLimits.Memory > 0 {
-		m.weakRefPool = weakref.NewDefaultPool()
+		m.weakRefPool = luagc.NewDefaultPool()
 		m.gcPolicy = IsolateGCPolicy
 	} else {
 		m.weakRefPool = parent.weakRefPool
